# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.Delivery:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      region:
        type: string
      zip:
        type: string
    required:
    - address
    - city
    - email
    - name
    - phone
    - region
    - zip
    type: object
  model.Order:
    properties:
      customer_id:
        type: string
      date_created:
        type: string
      delivery:
        $ref: '#/definitions/model.Delivery'
      delivery_service:
        type: string
      entry:
        type: string
      internal_signature:
        type: string
      items:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      locale:
        type: string
      oof_shard:
        type: string
      order_uid:
        type: string
      payment:
        $ref: '#/definitions/model.Payment'
      shardkey:
        type: string
      sm_id:
        type: integer
      track_number:
        type: string
    required:
    - customer_id
    - date_created
    - delivery
    - delivery_service
    - entry
    - items
    - locale
    - oof_shard
    - order_uid
    - payment
    - shardkey
    - sm_id
    - track_number
    type: object
  model.OrderItem:
    properties:
      brand:
        type: string
      chrt_id:
        type: integer
      name:
        type: string
      nm_id:
        type: integer
      price:
        type: integer
      rid:
        type: string
      sale:
        type: integer
      size:
        type: string
      status:
        type: integer
      total_price:
        type: integer
      track_number:
        type: string
    required:
    - brand
    - chrt_id
    - name
    - nm_id
    - price
    - rid
    - sale
    - size
    - status
    - total_price
    - track_number
    type: object
  model.Payment:
    properties:
      amount:
        type: integer
      bank:
        type: string
      currency:
        type: string
      custom_fee:
        type: integer
      delivery_cost:
        type: integer
      goods_total:
        type: integer
      payment_dt:
        type: integer
      provider:
        type: string
      request_id:
        type: string
      transaction:
        type: string
    required:
    - amount
    - bank
    - currency
    - custom_fee
    - delivery_cost
    - goods_total
    - payment_dt
    - provider
    - transaction
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /order:
    post:
      consumes:
      - application/json
      description: Create a new order with the given details
      parameters:
      - description: Order Request
        in: body
        name: order
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Invalid request body or invalid order data
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Order already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new order
      tags:
      - order
  /order/{uuid}:
    get:
      consumes:
      - application/json
      description: Get details of an order by its UUID
      parameters:
      - description: Order UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Order'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get an order by UUID
      tags:
      - order
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"